api-v2/services/common/notification-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
httpx==0.25.2
motor==3.6.0
python-dotenv==1.0.0
aiokafka==0.12.0
